backend/requirements.txt
fastapi
pydantic
uvicorn
gitpython
python-slugify
markdown
python-frontmatter
